MSVC compiler builds on Windows

+++ src/core/mpz.hpp
@@ -932,7 +932,7 @@ public:
         m_limbs.resize((used < 1)? 1 : used);
         used = mpz_core<T>::abs_sub(m_limbs.data(), in1.get_limbs().data(), used, in2);
         m_limbs.resize((used < 0)? -used : used);
-        return used;
+        return static_cast<int32_t>(used);
     }
 
     /// Add an unsigned integer to an mpz object
